fix: 自然排序修复 + 代码清理
- 修复视频/文件夹列表名称排序为自然排序（1,2,3,10,11 而非 1,10,11,2,3）
- 修复树状视图刷新导致面包屑无限嵌套（1>2>3>3>3...）
- 修复树状视图四个代码路径跳过排序的问题
- 提取 NaturalOrderComparator 公共工具类，消除6处重复
- 提取 FormatUtils.formatDuration，统一4处重复 formatDuration/formatFileSize
- 提取 SortOption 公共组件，消除3处重复
- 统一 EmptyState 为 CommonStates 公共组件
- 删除 UI 层死代码（sortFolders/sortVideos/filterVideos/naturalComparator/compareNatural）
