+++ app/src/main/java/com/fam4k007/videoplayer/utils/FormatUtils.kt
@@ -3,6 +3,7 @@ package com.fam4k007.videoplayer.utils
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
+import java.util.concurrent.TimeUnit
 
 /**
  * 统一格式化工具类
@@ -57,6 +58,22 @@ object FormatUtils {
             bytes < 1024 * 1024 -> String.format(Locale.getDefault(), "%.2f KB", bytes / 1024.0)
             bytes < 1024 * 1024 * 1024 -> String.format(Locale.getDefault(), "%.2f MB", bytes / (1024.0 * 1024.0))
             else -> String.format(Locale.getDefault(), "%.2f GB", bytes / (1024.0 * 1024.0 * 1024.0))
+        }
+    }
+
+    /**
+     * 格式化时长（毫秒 → HH:MM:SS 或 MM:SS）
+     * @param milliseconds 毫秒数
+     * @return 格式化后的时长字符串，例如 "01:23:45" 或 "23:45"
+     */
+    fun formatDuration(milliseconds: Long): String {
+        val hours = TimeUnit.MILLISECONDS.toHours(milliseconds)
+        val minutes = TimeUnit.MILLISECONDS.toMinutes(milliseconds) % 60
+        val seconds = TimeUnit.MILLISECONDS.toSeconds(milliseconds) % 60
+        return if (hours > 0) {
+            String.format(Locale.getDefault(), "%02d:%02d:%02d", hours, minutes, seconds)
+        } else {
+            String.format(Locale.getDefault(), "%02d:%02d", minutes, seconds)
         }
     }
